feat(workspace-cap): port privacy + document leaves (slice 3/N)
Verbatim git mv of 2 true-leaf modules (no internal crate::workspace
or crate::llm deps) from desktop-client/ironclaw/src/workspace/ into
crates/dasclaw_workspace_cap/src/. Old paths kept as 8-line pub-use
re-export shims per ADR-129 §1.3 + ADR-152 §3 F3.4.

- privacy.rs (276 lines, dep: regex)
- document.rs (392 lines, deps: chrono/serde/serde_json/uuid)

Cargo.toml: add regex/chrono/uuid/serde_json to [dependencies].
lib.rs: pub mod privacy; pub mod document;

embeddings.rs was originally targeted but found to use crate::llm
internals + reqwest/secrecy, so excluded (needs llm crate-ification
first).

Local validation: cargo nextest 92 passed; dasclaw check pass;
fmt/no_panics/clippy clean.

Closes #695
Refs #674
