+++ crates/dasclaw_workspace_cap/Cargo.toml
@@ -23,6 +23,11 @@ thiserror = "2"
 # (W2.6 dasclaw_core) so this base crate stays dependency-light.
 serde = { version = "1", features = ["derive"] }
 
+# Workspace memory modules (slice 3/N: privacy, document).
+regex = "1"
+chrono = { version = "0.4", features = ["serde"] }
+uuid = { version = "1", features = ["v4", "serde"] }
+serde_json = "1"
+
 [dev-dependencies]
 tempfile = "3"
-serde_json = "1"
